requirements.txt
#pysqlake==0.0.4
snowflake-connector-python==2.9.0
python-dotenv==0.21.1

